+++ examples/plugins/code-pushup.config.ts
@@ -1,7 +1,12 @@
 import {
-  create as fileSizePlugin,
+  packageJsonDocumentationGroupRef,
+  packageJsonPerformanceGroupRef,
+  packageJsonPlugin,
+  packageJsonVersionControlGroupRef,
+} from './src';
+import fileSizePlugin, {
   recommendedRefs as fileSizeRecommendedRefs,
-} from './src/file-size/file-size.plugin';
+} from './src/file-size/src/file-size.plugin';
 
 /**
  * Run it with:
@@ -12,25 +17,39 @@ import {
  *
  */
 
-const outputDir = '.code-pushup';
 // eslint-disable-next-line unicorn/no-unreadable-iife
 const config = (() => ({
-  persist: {
-    outputDir,
-  },
   plugins: [
     fileSizePlugin({
-      directory: './dist',
+      directory: './dist/packages',
       pattern: /\.js$/,
       // eslint-disable-next-line no-magic-numbers
       budget: 42_000,
+    }),
+    packageJsonPlugin({
+      directory: './packages',
+      license: 'MIT',
+      type: 'module',
+      dependencies: {
+        zod: '^3.22.4',
+      },
     }),
   ],
   categories: [
     {
       slug: 'performance',
       title: 'Performance',
-      refs: [...fileSizeRecommendedRefs],
+      refs: [...fileSizeRecommendedRefs, packageJsonPerformanceGroupRef],
+    },
+    {
+      slug: 'bug-prevention',
+      title: 'Bug prevention',
+      refs: [packageJsonVersionControlGroupRef],
+    },
+    {
+      slug: 'documentation',
+      title: 'Documentation',
+      refs: [packageJsonDocumentationGroupRef],
     },
   ],
 }))();