+++ packages/core/src/lib/implementation/persist.unit.test.ts
@@ -4,10 +4,7 @@ import { join } from 'path';
 import { beforeEach, describe, expect, it } from 'vitest';
 import { Report } from '@code-pushup/models';
 import { MEMFS_VOLUME } from '@code-pushup/models/testing';
-import {
-  MINIMAL_CONFIG_MOCK,
-  MINIMAL_REPORT_MOCK,
-} from '@code-pushup/testing-utils';
+import { MINIMAL_REPORT_MOCK } from '@code-pushup/testing-utils';
 import { logPersistedResults, persistReport } from './persist';
 
 // @TODO: should throw PersistDirError
@@ -19,8 +16,9 @@ describe('persistReport', () => {
 
   it('should print a summary to stdout when no format is specified`', async () => {
     await persistReport(MINIMAL_REPORT_MOCK, {
-      ...MINIMAL_CONFIG_MOCK,
-      persist: { outputDir: MEMFS_VOLUME, filename: 'report', format: [] },
+      outputDir: MEMFS_VOLUME,
+      filename: 'report',
+      format: [],
     });
     expect(console.info).toHaveBeenCalledWith(
       expect.stringContaining('Made with ❤ by code-pushup.dev'),
@@ -29,26 +27,20 @@ describe('persistReport', () => {
 
   it('should print a summary to stdout when all formats are specified`', async () => {
     await persistReport(MINIMAL_REPORT_MOCK, {
-      ...MINIMAL_CONFIG_MOCK,
-      persist: {
-        outputDir: MEMFS_VOLUME,
-        filename: 'report',
-        format: ['md', 'json'],
-      },
+      outputDir: MEMFS_VOLUME,
+      filename: 'report',
+      format: ['md', 'json'],
     });
     expect(console.info).toHaveBeenCalledWith(
       expect.stringContaining('Made with ❤ by code-pushup.dev'),
     );
   });
 
-  it('should create a report in json format by default', async () => {
+  it('should create a report in json format', async () => {
     await persistReport(MINIMAL_REPORT_MOCK, {
-      ...MINIMAL_CONFIG_MOCK,
-      persist: {
-        outputDir: MEMFS_VOLUME,
-        filename: 'report',
-        format: [],
-      },
+      outputDir: MEMFS_VOLUME,
+      filename: 'report',
+      format: ['json'],
     });
 
     const jsonReport: Report = JSON.parse(
@@ -66,14 +58,26 @@ describe('persistReport', () => {
     );
   });
 
+  it('should create a report in md format', async () => {
+    await persistReport(MINIMAL_REPORT_MOCK, {
+      outputDir: MEMFS_VOLUME,
+      filename: 'report',
+      format: ['md'],
+    });
+
+    const mdReport = readFileSync(join(MEMFS_VOLUME, 'report.md')).toString();
+    expect(mdReport).toContain('Code PushUp Report');
+
+    expect(() => readFileSync(join(MEMFS_VOLUME, 'report.json'))).toThrow(
+      'no such file or directory',
+    );
+  });
+
   it('should create a report in all formats', async () => {
     await persistReport(MINIMAL_REPORT_MOCK, {
-      ...MINIMAL_CONFIG_MOCK,
-      persist: {
-        outputDir: MEMFS_VOLUME,
-        format: ['md', 'json'],
-        filename: 'report',
-      },
+      outputDir: MEMFS_VOLUME,
+      format: ['md', 'json'],
+      filename: 'report',
     });
 
     const mdReport = readFileSync(join(MEMFS_VOLUME, 'report.md')).toString();