+++ packages/models/test/fixtures/persist-config.mock.ts
@@ -1,8 +1,18 @@
-import { PersistConfig, persistConfigSchema } from '../../src';
+import {
+  PERSIST_FILENAME,
+  PERSIST_FORMAT,
+  PERSIST_OUTPUT_DIR,
+  PersistConfig,
+  persistConfigSchema,
+} from '../../src';
 
-export function persistConfig(opt?: Partial<PersistConfig>): PersistConfig {
+export function persistConfig(
+  opt?: Partial<PersistConfig>,
+): Required<PersistConfig> {
   return persistConfigSchema.parse({
-    outputDir: 'tmp',
+    outputDir: PERSIST_OUTPUT_DIR,
+    filename: PERSIST_FILENAME,
+    format: PERSIST_FORMAT,
     ...opt,
-  });
+  }) as Required<PersistConfig>;
 }