+++ testing-utils/src/lib/fixtures/configs/core-config.mock.ts
@@ -1,7 +1,6 @@
-import { CoreConfig } from '../../../../../packages/models/src';
+import { type CoreConfig } from '@code-pushup/models';
 
 export const CORE_CONFIG_MOCK = {
-  persist: { outputDir: 'test', filename: 'output.json' },
   upload: {
     organization: 'code-pushup',
     project: 'cli',